% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_aggregate_matrix.R
\name{lvdr_aggregate_matrix}
\alias{lvdr_aggregate_matrix}
\title{Query CELEX database}
\usage{
lvdr_aggregate_matrix(tfm, group, fun = "sum")
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
