% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_calc_ppmi.R
\name{lvdr_build_sparse_ppmi}
\alias{lvdr_build_sparse_ppmi}
\alias{lvdr_calc_ppmi}
\title{Query CELEX database}
\usage{
lvdr_calc_ppmi(pmat, make_symmetric = TRUE)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
