% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_extract_network.R
\name{lvdr_extract_network}
\alias{lvdr_extract_network}
\title{Query CELEX database}
\usage{
lvdr_extract_network(y, search, min_val)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
