% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_get_associates.R
\name{lvdr_get_associates}
\alias{lvdr_get_associates}
\title{Query CELEX database}
\usage{
lvdr_get_associates(cue, toupper = FALSE)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
