% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_get_closest.R
\name{lvdr_extract_network}
\alias{lvdr_extract_network}
\alias{lvdr_get_closest}
\title{Query CELEX database}
\usage{
lvdr_get_closest(tfm, target = NULL, n = NULL, include_targ = FALSE)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
