% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_get_family.R
\name{lvdr_get_family}
\alias{lvdr_get_family}
\title{Query CELEX database}
\usage{
lvdr_get_family(form, type = "word", multiword = FALSE, toupper = FALSE,
  include_pos = FALSE)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
