% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvdr_quick_cosine.R
\name{lvdr_quick_cosine}
\alias{lvdr_quick_cosine}
\title{Query CELEX database}
\usage{
lvdr_quick_cosine(tfm, target, n = 10, vocab = NULL)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
