% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LexvarFunctions.R
\name{LexvarFunctions}
\alias{LexvarFunctions}
\alias{getFamily}
\title{Query CELEX database}
\usage{
getFamily(form, type)
}
\arguments{
\item{form}{A word or affix}

\item{type}{A part-of-speach (of universal variety) or "PRE"|"SUF"}
}
\value{
A data.frame
}
\description{
These functions make life easier when querying the behavioral datasets.
}
